odin: Move mockRDSClient to internal package
Make it into an internal package and reuse it for testing in both
`pkg/odin` and `pkg/odin_test`.
