+++ pkg/odin/odin_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/rds"
 
 	"github.com/poka-yoke/spaceflight/internal/test_case"
+	"github.com/poka-yoke/spaceflight/internal/test/mockRDSClient"
 	"github.com/poka-yoke/spaceflight/pkg/odin"
 )
 
@@ -57,7 +58,7 @@ var getLastSnapshotCases = []getLastSnapshotCase{
 }
 
 func TestGetLastSnapshot(t *testing.T) {
-	svc := newMockRDSClient()
+	svc := mockrdsclient.NewMockRDSClient()
 	for _, test := range getLastSnapshotCases {
 		t.Run(
 			test.name,
@@ -190,7 +191,7 @@ func TestListSnapshots(t *testing.T) {
 		t.Run(
 			test.name,
 			func(t *testing.T) {
-				svc := newMockRDSClient()
+				svc := mockrdsclient.NewMockRDSClient()
 				svc.AddSnapshots(test.snapshots)
 				actual, err := odin.ListSnapshots(
 					test.instanceID,
